requirements.txt
beautifulsoup4~=4.13.3
requests~=2.32.3
loguru~=0.7.3
PyYAML~=6.0.2
PyPDF2~=3.0.1
reportlab~=4.3.1
weasyprint~=64.1
pathlib~=1.0.1
PyMuPDF~=1.25.3